requirements.txt
alabaster==0.7.11
apipkg==1.5
appnope==0.1.0
argh==0.26.2
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
bumpversion==0.5.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.3.0
docutils==0.14
execnet==1.5.0
flake8==3.5.0
idna==2.7
imagesize==1.1.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
networkx==2.1
packaging==17.1
parso==0.3.1
pathtools==0.1.2
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
pkginfo==1.4.2
pluggy==0.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.8.0
pytest-cache==1.0
pytest-cov==2.6.0
pytest-pep8==1.0.6
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.1
sphinxcontrib-websupport==1.1.0
tox==2.9.1
tqdm==4.25.0
traitlets==4.3.2
twine==1.10.0
urllib3==1.23
virtualenv==16.0.0
watchdog==0.8.3
wcwidth==0.1.7
Whoosh==2.7.4
